realtek-poe: move blob_buf out of global scope
the `blob_buf` is only used in ubus_poe_info_cb(). Since it is also
initialized in the same function, it could be reasoned that it should
be a stack variable. However, placing it on the stack seems to cause
segmentation faults, as observed when calling `ubus call poe info`.
Presumably, this is why `blob_buf` was originally made global.

To take `blob_buf` out of global scope, without casuing issues, move
it to the poe context.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>